% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_comparisons.R
\name{jags_comparisons}
\alias{jags_comparisons}
\title{Compare JAGS models}
\usage{
jags_comparisons(jags_outputs)
}
\arguments{
\item{jags_outputs}{a list of the jags outputs to compare}
}
\value{
a dataframe with the model names as rows and columns with the DIC and pV scores
}
\description{
A function to compare the DIC and pV of multiple models run through JAGS
}
