% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_graphs.R
\name{regression_graphs}
\alias{regression_graphs}
\title{Create regression graphs with confidence intervals}
\usage{
regression_graphs(
  jags_output,
  dataframe,
  factor_column_name,
  variable_columns,
  factor_names,
  response_variable_column,
  jags_intercept_names,
  jags_slopes_names
)
}
\arguments{
\item{jags_output}{JAGS output file from the model}

\item{dataframe}{name of the dataframe containing the raw data input to the model}

\item{factor_column_name}{a string of the column name with the factors in dataframe}

\item{variable_columns}{a vector of strings of the column names that contain the variables in dataframe}

\item{factor_names}{a vector of strings with the names of the factors. This is for titling the graph and must be the same order as they are in dataframe}

\item{response_variable_column}{a string of the response variable column in dataframe}

\item{jags_intercept_names}{string of the name of the intercept variable in the JAGS file}

\item{jags_slope_names}{a vector of strings of the names of the slopes within the JAGS file}
}
\value{
a group of graphs with the regressions and 95\% confidence intervals of the JAGS model
}
\description{
A function to create graphs of the raw data with the modeled line and confidence intervals
}
