% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_graphs.R
\name{variable_selection_graphs}
\alias{variable_selection_graphs}
\title{Create variable selection graphs}
\usage{
variable_selection_graphs(jags_output, factor_names)
}
\arguments{
\item{jags_output}{a JAGS file of the model run}

\item{factor_names}{a vector of strings of the factor names being tested, must be the same order as the input data to the JAGS model. This is for titling purposes}
}
\value{
a group of variable selection graphs
}
\description{
A function to make variable selection graphs after running a lassoo to select variables
}
